Corrected filename. Removed references to v2
There is an error where a dash is used instead of an underline in my_overlay.

I removed references to v2 as it just confused me.  These are optional changes, of course :)

There is also something wrong although I can't identify it.  In line 56, it states "second example", however I'm not sure what that means as I only see one example.  Also in line 69, my_overlay/ is used.  Is that correct?  Or should it be my_overlay without the slash?

Finally, the mkfs.ext3 command shows a warning and a prompt when I run it.  Maybe that should be documented?
mkfs.ext3 my_overlay
500+0 records in
500+0 records out
524288000 bytes (524 MB) copied, 1.4015 s, 374 MB/s
mke2fs 1.42.9 (28-Dec-2013)
my_overlay is not a block special device.
Proceed anyway? (y,n) 

OS is CentOS 7